move read_file related tests to separate test file
